# Cache supplier for invocation rather than method resolving