Currently cache contains suppliers which then should resolve methods and then do invoke. In fact it is possible to keep result of resolving in cache. The PR does this

as a result it speeds up some benchmarks 
before
```
Benchmark                            Mode  Cnt     Score    Error   Units
JmhTest._optionsExpression          thrpt    5  1610.219 ± 24.256  ops/ms
JmhTest.oneMethodCallExpression     thrpt    5  1273.391 ± 33.416  ops/ms
JmhTest.threeMethodsCallExpression  thrpt    5   407.451 ± 35.170  ops/ms
``` 

after
```
Benchmark                            Mode  Cnt     Score    Error   Units
JmhTest._optionsExpression          thrpt    5  4713.630 ± 48.540  ops/ms
JmhTest.oneMethodCallExpression     thrpt    5  2277.320 ± 30.705  ops/ms
JmhTest.threeMethodsCallExpression  thrpt    5   552.462 ±  5.566  ops/ms
```

benchmark could be found at #25